Seta api authorization

+++ seta-api/apis/corpus/corpus.py
@@ -2,12 +2,13 @@ import json
 from flask import current_app as app
 from flask import jsonify, request
 from flask_restx import Namespace, Resource, abort, fields
-from infrastructure.ApiLogicError import ApiLogicError
-from infrastructure.auth_validator import auth_validator
+from infrastructure.ApiLogicError import ApiLogicError, ForbiddenResourceError
+from infrastructure.auth_validator import auth_validator, validate_view_permissions, validate_add_permission, validate_delete_permission
 from infrastructure.helpers import is_field_in_doc
 
 from .corpus_logic import corpus, delete_doc, docbyid, insert_doc
 from .variables import corpus_parser, keywords, metadata, other
+from http import HTTPStatus
 
 corpus_api = Namespace('seta-api-corpus', description='Corpus')
 
@@ -22,24 +23,38 @@ class Corpus(Resource):
             security='apikey')
     def get(self, id):
         try:
-            return docbyid(id, current_app=app)
+            doc = docbyid(id, current_app=app)
+            
+            validate_view_permissions(sources=[doc.get("source", None)])
+            
+            return doc
         except ApiLogicError as aex:
             abort(404, str(aex))
+        except ForbiddenResourceError as fre:
+            abort(HTTPStatus.FORBIDDEN, fre.message)
         except:
             app.logger.exception("Corpus->get")
             abort(500, "Internal server error")
 
-    @auth_validator("Administrator")
+    @auth_validator()
     @corpus_api.doc(description='Given the elasticsearch unique _id, the relative document is deleted.',
             params={'id': 'Delete the document with the specified _id'},
             responses={200: 'Success', 404: 'Not Found Error'},
             security='apikey')
     def delete(self, id):
         try:
+            doc = docbyid(id, current_app=app)            
+            resource_id = doc.get("source", None)
+            
+            if not validate_delete_permission(resource_id):
+                raise ForbiddenResourceError(resource_id=resource_id, message="User does not have delete permission for the resource")
+            
             delete_doc(id, current_app=app)
             return jsonify({"deleted document id": id})
         except ApiLogicError as aex:
             abort(404, str(aex))
+        except ForbiddenResourceError as fre:
+            abort(HTTPStatus.FORBIDDEN, fre.message)
         except:
             app.logger.exception("Corpus->delete")
             abort(500, "Internal server error")
@@ -115,7 +130,18 @@ class CorpusQuery(Resource):
     @corpus_api.expect(query_corpus_post_data)
     def post(self):
         args = request.get_json(force=True)
-        app.logger.debug(str(args))        
+        app.logger.debug(str(args))
+        
+        #validate resource_permissions.view
+        sources = is_field_in_doc(args, 'source')
+        try:
+            view_resources = validate_view_permissions(sources)
+            
+            #restrict query only to view_resources
+            if sources is None:
+                args["source"] = view_resources
+        except ForbiddenResourceError as fre:
+            abort(403, fre.message)
         
         if is_field_in_doc(args, 'term') or is_field_in_doc(args, 'semantic_sort_id') \
                 or is_field_in_doc(args, 'semantic_sort_id_list')\
@@ -155,12 +181,17 @@ class CorpusQuery(Resource):
                 app.logger.exception("CorpusQuery->post")
                 abort(500, "Internal server error")
 
-    @auth_validator("Administrator")
+    @auth_validator()
     @corpus_api.doc(description='Put a document into corpus index.',
             security='apikey')
     @corpus_api.expect(corpus_put_data)
     def put(self):
         args = request.get_json(force=True)
+        
+        source = is_field_in_doc(args, 'source')
+        if not validate_add_permission(source):
+            abort(HTTPStatus.FORBIDDEN, "User does not have add document permission for the resource")
+        
         doc_id = insert_doc(args, current_app=app)
         return jsonify({"document_id": doc_id})
 
@@ -195,7 +226,19 @@ class CorpusQuery(Resource):
                     'aggs': 'field to be aggregated, allowed fields are: "source", "eurovoc_concept"'},
             security='apikey')
     def get(self):
-        args = corpus_parser.parse_args()
+        args = corpus_parser.parse_args()        
+        
+        #validate resource_permissions.view
+        sources = is_field_in_doc(args, 'source')
+        try:
+            view_resources = validate_view_permissions(sources)
+            
+            #restrict query only to view_resources
+            if sources is None:
+                args["source"] = view_resources
+        except ForbiddenResourceError as fre:
+            abort(403, fre.message)       
+        
         if args['term'] or args['semantic_sort_id'] or args['semantic_sort_id_list'] or args['aggs'] or args['source']:
             try:
                 documents = corpus(args['term'], args['n_docs'], args['from_doc'], args['source'], args['collection'],
